+++ preset/ie11.js
@@ -1,0 +1,5 @@
+/* eslint-disable node/no-missing-require, @typescript-eslint/no-unsafe-assignment */
+
+module.exports = {
+  ...require('../lib/cjs/client/addon/preset.ie11'),
+};