+++ src/server/selenium/browser.ts
@@ -21,7 +21,7 @@ import {
 } from '../../types';
 import { colors, logger } from '../logger';
 import { emitStoriesMessage, subscribeOn } from '../messages';
-import { importStorybookCoreEvents, isStorybookVersionLessThan } from '../storybook/helpers';
+import { importStorybookCoreEvents } from '../storybook/helpers';
 import { isShuttingDown, LOCALHOST_REGEXP, runSequence } from '../utils';
 
 type ElementRect = {
@@ -64,7 +64,7 @@ function getSessionData(grid: string, sessionId = ''): Promise<Record<string, un
       res.on('data', (chunk) => (data += chunk));
       res.on('end', () => {
         try {
-          resolve(JSON.parse(data));
+          resolve(JSON.parse(data) as Record<string, unknown>);
         } catch (error) {
           reject(
             new Error(
@@ -147,16 +147,6 @@ function getUrlChecker(browser: WebDriver): (url: string) => Promise<boolean> {
 }
 
 async function waitForStorybook(browser: WebDriver): Promise<void> {
-  // NOTE: Storybook 5.x doesn't have the `last` method
-  if (isStorybookVersionLessThan(6)) {
-    browserLogger.debug('Waiting for `load` event to make sure that storybook is initiated');
-    return browser.executeAsyncScript(function (callback: () => void): void {
-      if (document.readyState == 'complete') return callback();
-      window.addEventListener('load', function () {
-        callback();
-      });
-    });
-  }
   browserLogger.debug('Waiting for `setStories` event to make sure that storybook is initiated');
   let wait = true;
   let isTimeout = false;
@@ -421,9 +411,7 @@ async function selectStory(
   waitForReady = false,
 ): Promise<boolean> {
   browserLogger.debug(`Triggering 'SetCurrentStory' event with storyId ${chalk.magenta(id)}`);
-  const [errorMessage, isCaptureCalled = false] = await browser.executeAsyncScript<
-    [error?: string | null, isCaptureCalled?: boolean]
-  >(
+  const result = await browser.executeAsyncScript<[error?: string | null, isCaptureCalled?: boolean] | null>(
     function (
       id: string,
       kind: string,
@@ -436,23 +424,22 @@ async function selectStory(
           "Creevey can't switch story. This may happened if forget to add `creevey` addon to your storybook config, or storybook not loaded in browser due syntax error.",
         ]);
       }
-      window.__CREEVEY_SELECT_STORY__(id, kind, name, shouldWaitForReady, callback);
+      void window.__CREEVEY_SELECT_STORY__(id, kind, name, shouldWaitForReady, callback);
     },
     id,
     kind,
     name,
     waitForReady,
   );
+
+  const [errorMessage, isCaptureCalled = false] = result ?? [];
+
   if (errorMessage) throw new Error(errorMessage);
 
   return isCaptureCalled;
 }
 
 export async function updateStorybookGlobals(browser: WebDriver, globals: StorybookGlobals): Promise<void> {
-  if (isStorybookVersionLessThan(6)) {
-    browserLogger.warn('Globals are not supported by Storybook versions less than 6');
-    return;
-  }
   browserLogger.debug('Applying storybook globals');
   await browser.executeScript(function (globals: StorybookGlobals) {
     window.__CREEVEY_UPDATE_GLOBALS__(globals);
@@ -499,7 +486,8 @@ async function resolveCreeveyHost(browser: WebDriver, port: number): Promise<str
     function (hosts: string[], port: number, callback: (host?: string | null) => void) {
       void Promise.all(
         hosts.map(function (host) {
-          return fetch(`http://${host}:${port}/ping`)
+          // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
+          return fetch('http://' + host + ':' + port + '/ping')
             .then(function (response) {
               return response.text();
             })
@@ -564,8 +552,7 @@ export async function getBrowser(config: Config, name: string): Promise<WebDrive
   const realAddress = address;
 
   // TODO Define some capabilities explicitly and define typings
-  const capabilities = new Capabilities(userCapabilities);
-  capabilities.setPageLoadStrategy(PageLoadStrategy.NONE);
+  const capabilities = new Capabilities({ ...userCapabilities, pageLoadStrategy: PageLoadStrategy.NONE });
 
   subscribeOn('shutdown', () => {
     browser?.quit().finally(() =>
@@ -603,7 +590,7 @@ export async function getBrowser(config: Config, name: string): Promise<WebDrive
 
     prefix.apply(browserLogger, {
       format(level) {
-        const levelColor = colors[level.toUpperCase()];
+        const levelColor = colors[level.toUpperCase() as keyof typeof colors];
         return `[${name}:${chalk.gray(sessionId)}] ${levelColor(level)} =>`;
       },
     });