+++ src/server/stories.ts
@@ -15,12 +15,13 @@ import type {
 } from '../types';
 import { isDefined, isFunction, isObject } from '../types';
 import { shouldSkip, removeProps } from './utils';
-import { isStorybookVersionLessThan } from './storybook/helpers';
 
 function storyTestFabric(delay?: number, testFn?: CreeveyTestFunction) {
   return async function storyTest(this: Context) {
     delay ? await new Promise((resolve) => setTimeout(resolve, delay)) : void 0;
-    await (testFn?.call(this) ?? this.screenshots.length > 0
+    await (testFn
+      ? testFn.call(this)
+      : this.screenshots.length > 0
       ? this.expect(
           this.screenshots.reduce(
             (screenshots, { imageName, screenshot }, index) => ({
@@ -130,18 +131,16 @@ export async function loadTestsFromStories(
 export function saveStoriesJson(storiesData: SetStoriesData, extract: string | boolean): void {
   const outputDir = typeof extract == 'boolean' ? 'storybook-static' : extract;
 
-  if (!isStorybookVersionLessThan(6)) {
-    // NOTE Copy-pasted from Storybook's `getStoriesJsonData` method
-    const allowed = ['fileName', 'docsOnly', 'framework', '__id', '__isArgsStory'];
-    storiesData.globalParameters = pick(storiesData.globalParameters, allowed);
-    // @ts-expect-error ignore error
-    storiesData.kindParameters = mapValues(storiesData.kindParameters, (v) => pick(v, allowed));
-    // @ts-expect-error ignore error
-    storiesData.stories = mapValues(storiesData.stories, (v) => ({
-      ...pick(v, ['id', 'name', 'kind', 'story']),
-      parameters: pick(v.parameters, allowed),
-    }));
-  }
+  // NOTE Copy-pasted from Storybook's `getStoriesJsonData` method
+  const allowed = ['fileName', 'docsOnly', 'framework', '__id', '__isArgsStory'];
+  storiesData.globalParameters = pick(storiesData.globalParameters, allowed);
+  // @ts-expect-error ignore error
+  storiesData.kindParameters = mapValues(storiesData.kindParameters, (v) => pick(v, allowed));
+  // @ts-expect-error ignore error
+  storiesData.stories = mapValues(storiesData.stories, (v) => ({
+    ...pick(v, ['id', 'name', 'kind', 'story']),
+    parameters: pick(v.parameters, allowed),
+  }));
 
   // TODO Fix args stories
   removeProps(storiesData ?? {}, ['stories', () => true, 'parameters', '__isArgsStory']);