+++ tests/e2e/storybook.fixtures/sb-5.0-require/.storybook/config.js
@@ -1,7 +1,0 @@
-import { configure, addDecorator, addParameters } from '@storybook/react';
-import { withCreevey } from 'creevey'
-
-addDecorator(withCreevey())
-addParameters({ creevey: { captureElement: 'root' }})
-
-configure(() => [require('../stories/KindA.stories')], module);